Added attack speed update

+++ src/main/java/me/despical/battleacademy/api/StatsStorage.java
@@ -29,7 +29,7 @@ public class StatsStorage {
 		return SortUtils.sortByValue(stats);
 	}
 
-	public static int getUserStats(final Player player, final StatisticType statisticType) {
+	public static double getUserStats(final Player player, final StatisticType statisticType) {
 		return plugin.getUserManager().getUser(player).getStat(statisticType);
 	}
